feat: Support creation of color maps through SLD filters
Inside styler add a button to generate filters to categorize selected variable into 11 bins and set
a color for symbols from various colormaps supported by colormap library.

BREAKING CHANGE: npm i colormaps

+++ projects/hslayers/src/common/layers/hs.source.interpolated.ts
@@ -190,7 +190,7 @@ export class InterpolatedSource extends IDW {
    */
   normalizeWeight(weight: string): void {
     const features = this.featureCache.getFeatures();
-    const weightValues = features.map((f) => parseInt(f.get(weight)));
+    const weightValues = features.map((f) => parseFloat(f.get(weight)));
     const min = Math.min(...weightValues);
     const max = Math.max(...weightValues);
